[mvsData] Merge imageIO.hpp to image/io.hpp
