+++ src/aliceVision/fuseCut/Fuser.cpp
@@ -14,7 +14,7 @@
 #include <aliceVision/mvsData/Stat3d.hpp>
 #include <aliceVision/mvsUtils/common.hpp>
 #include <aliceVision/mvsUtils/fileIO.hpp>
-#include <aliceVision/mvsData/imageIO.hpp>
+#include <aliceVision/image/io.hpp>
 #include <aliceVision/image/imageAlgo.hpp>
 
 #include <boost/filesystem.hpp>
@@ -47,9 +47,9 @@ unsigned long computeNumberOfAllPoints(const mvsUtils::MultiViewParams& mp, int 
 
             ALICEVISION_LOG_WARNING("Can't find or invalid 'nbDepthValues' metadata in '" << filename << "'. Recompute the number of valid values.");
 
-            imageIO::readImage(mvsUtils::getFileNameFromIndex(mp, rc, mvsUtils::EFileType::depthMap, scale),
-                               width, height, depthMap.getDataWritable(),
-                               image::EImageColorSpace::NO_CONVERSION);
+            image::readImage(mvsUtils::getFileNameFromIndex(mp, rc, mvsUtils::EFileType::depthMap, scale),
+                             width, height, depthMap.getDataWritable(),
+                             image::EImageColorSpace::NO_CONVERSION);
             // no need to transpose for this operation
             for(int i = 0; i < sizeOfStaticVector<float>(&depthMap); ++i)
                 nbDepthValues += static_cast<unsigned long>(depthMap[i] > 0.0f);
@@ -166,12 +166,12 @@ bool Fuser::filterGroupsRC(int rc, float pixToleranceFactor, int pixSizeBall, in
     {
         int width, height;
 
-        imageIO::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::depthMap, 1),
-                           width, height, depthMap.getDataWritable(),
-                           image::EImageColorSpace::NO_CONVERSION);
-        imageIO::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::simMap, 1),
-                           width, height, simMap.getDataWritable(),
-                           image::EImageColorSpace::NO_CONVERSION);
+        image::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::depthMap, 1),
+                         width, height, depthMap.getDataWritable(),
+                         image::EImageColorSpace::NO_CONVERSION);
+        image::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::simMap, 1),
+                         width, height, simMap.getDataWritable(),
+                         image::EImageColorSpace::NO_CONVERSION);
     }
 
     std::vector<unsigned char> numOfModalsMap(w * h, 0);
@@ -199,9 +199,9 @@ bool Fuser::filterGroupsRC(int rc, float pixToleranceFactor, int pixSizeBall, in
 
         {
             int width, height;
-            imageIO::readImage(getFileNameFromIndex(_mp, tc, mvsUtils::EFileType::depthMap, 1),
-                               width, height, tcdepthMap.getDataWritable(),
-                               image::EImageColorSpace::NO_CONVERSION);
+            image::readImage(getFileNameFromIndex(_mp, tc, mvsUtils::EFileType::depthMap, 1),
+                             width, height, tcdepthMap.getDataWritable(),
+                             image::EImageColorSpace::NO_CONVERSION);
         }
 
         if(!tcdepthMap.empty())
@@ -227,10 +227,9 @@ bool Fuser::filterGroupsRC(int rc, float pixToleranceFactor, int pixSizeBall, in
     }
 
     {
-        using namespace imageIO;
-        writeImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::nmodMap), w, h,
-                   numOfModalsMap, EImageQuality::LOSSLESS,
-                   OutputFileColorSpace(image::EImageColorSpace::NO_CONVERSION));
+        image::writeImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::nmodMap), w, h,
+                          numOfModalsMap, image::EImageQuality::LOSSLESS,
+                          image::OutputFileColorSpace(image::EImageColorSpace::NO_CONVERSION));
     }
 
     delete numOfPtsMap;
@@ -271,12 +270,12 @@ bool Fuser::filterDepthMapsRC(int rc, int minNumOfModals, int minNumOfModalsWSP2
     {
         int width, height;
 
-        imageIO::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::depthMap, 1),
-                           width, height, depthMap, image::EImageColorSpace::NO_CONVERSION);
-        imageIO::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::simMap, 1),
-                           width, height, simMap, image::EImageColorSpace::NO_CONVERSION);
-        imageIO::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::nmodMap),
-                           width, height, numOfModalsMap, image::EImageColorSpace::NO_CONVERSION);
+        image::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::depthMap, 1),
+                         width, height, depthMap, image::EImageColorSpace::NO_CONVERSION);
+        image::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::simMap, 1),
+                         width, height, simMap, image::EImageColorSpace::NO_CONVERSION);
+        image::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::nmodMap),
+                         width, height, numOfModalsMap, image::EImageColorSpace::NO_CONVERSION);
     }
 
     int nbDepthValues = 0;
@@ -330,13 +329,12 @@ bool Fuser::filterDepthMapsRC(int rc, int minNumOfModals, int minNumOfModalsWSP2
       metadata.push_back(oiio::ParamValue("AliceVision:P", oiio::TypeDesc(oiio::TypeDesc::DOUBLE, oiio::TypeDesc::MATRIX44), 1, matrixP.data()));
     }
 
-    using namespace imageIO;
-    writeImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::depthMap, 0), w, h, depthMap,
-               EImageQuality::LOSSLESS,
-               OutputFileColorSpace(image::EImageColorSpace::NO_CONVERSION), metadata);
-    writeImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::simMap, 0), w, h, simMap,
-               EImageQuality::OPTIMIZED,
-               OutputFileColorSpace(image::EImageColorSpace::NO_CONVERSION), metadata);
+    image::writeImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::depthMap, 0), w, h, depthMap,
+                      image::EImageQuality::LOSSLESS,
+                      image::OutputFileColorSpace(image::EImageColorSpace::NO_CONVERSION), metadata);
+    image::writeImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::simMap, 0), w, h, simMap,
+                      image::EImageQuality::OPTIMIZED,
+                      image::OutputFileColorSpace(image::EImageColorSpace::NO_CONVERSION), metadata);
 
     ALICEVISION_LOG_DEBUG(rc << " solved.");
     mvsUtils::printfElapsedTime(t1);
@@ -363,9 +361,9 @@ float Fuser::computeAveragePixelSizeInHexahedron(Point3d* hexah, int step, int s
 
         {
             int width, height;
-            imageIO::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::depthMap, scale),
-                               width, height, rcdepthMap.getDataWritable(),
-                               image::EImageColorSpace::NO_CONVERSION);
+            image::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::depthMap, scale),
+                             width, height, rcdepthMap.getDataWritable(),
+                             image::EImageColorSpace::NO_CONVERSION);
         }
 
         for(int y = 0; y < h; y++)
@@ -463,9 +461,9 @@ void Fuser::divideSpaceFromDepthMaps(Point3d* hexah, float& minPixSize)
         {
             int width, height;
 
-            imageIO::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::depthMap, scale),
-                               width, height, depthMap.getDataWritable(),
-                               image::EImageColorSpace::NO_CONVERSION);
+            image::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::depthMap, scale),
+                             width, height, depthMap.getDataWritable(),
+                             image::EImageColorSpace::NO_CONVERSION);
         }
 
         for(int i = 0; i < sizeOfStaticVector<float>(&depthMap); i += stepPts)
@@ -509,9 +507,9 @@ void Fuser::divideSpaceFromDepthMaps(Point3d* hexah, float& minPixSize)
         {
             int width, height;
 
-            imageIO::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::depthMap, scale),
-                               width, height, depthMap.getDataWritable(),
-                               image::EImageColorSpace::NO_CONVERSION);
+            image::readImage(getFileNameFromIndex(_mp, rc, mvsUtils::EFileType::depthMap, scale),
+                             width, height, depthMap.getDataWritable(),
+                             image::EImageColorSpace::NO_CONVERSION);
         }
 
         for(int i = 0; i < depthMap.size(); i += stepPts)
@@ -786,12 +784,12 @@ std::string generateTempPtsSimsFiles(const std::string& tmpDir, mvsUtils::MultiV
             {
                 int width, height;
 
-                imageIO::readImage(getFileNameFromIndex(mp, rc, mvsUtils::EFileType::depthMap, scale),
-                                   width, height, depthMap.getDataWritable(),
-                                   image::EImageColorSpace::NO_CONVERSION);
-                imageIO::readImage(getFileNameFromIndex(mp, rc, mvsUtils::EFileType::simMap, scale),
-                                   width, height, simMap.getDataWritable(),
-                                   image::EImageColorSpace::NO_CONVERSION);
+                image::readImage(getFileNameFromIndex(mp, rc, mvsUtils::EFileType::depthMap, scale),
+                                 width, height, depthMap.getDataWritable(),
+                                 image::EImageColorSpace::NO_CONVERSION);
+                image::readImage(getFileNameFromIndex(mp, rc, mvsUtils::EFileType::simMap, scale),
+                                 width, height, simMap.getDataWritable(),
+                                 image::EImageColorSpace::NO_CONVERSION);
             }
 
             if(addRandomNoise)
