+++ src/aliceVision/mesh/Texturing.hpp
@@ -7,7 +7,7 @@
 #pragma once
 
 #include <aliceVision/image/io.hpp>
-#include <aliceVision/mvsData/imageIO.hpp>
+#include <aliceVision/image/io.hpp>
 #include <aliceVision/mvsData/Point2d.hpp>
 #include <aliceVision/mvsData/Point3d.hpp>
 #include <aliceVision/mvsData/StaticVector.hpp>
